Extract and convert the 'Data Types' info callout markup into a reusable component
